refactor(operator): consolidate pod builders and tests into job_builder
Move createPodFromPackage, createInterruptPodForPackage, and
podMatchesPackage out of skyhook_controller.go into job_builder.go, and
relocate their specs from skyhook_controller_test.go into
job_builder_test.go so the builders and their tests live together.

Add coverage for the image-pull-secret-set path on both builders, the
graceful-shutdown to terminationGracePeriodSeconds mapping, and the
interrupt pod name/label/root-mount shape.
