+++ operator/internal/controller/job_builder.go
@@ -21,13 +21,17 @@ package controller
 import (
 	"fmt"
 	"math"
+	"reflect"
+	"sort"
 	"strconv"
+	"strings"
 	"time"
 
 	"github.com/NVIDIA/nodewright/operator/api/nodewright/v1alpha1"
 	"github.com/NVIDIA/nodewright/operator/internal/wrapper"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -163,4 +167,386 @@ func effectiveStageTimeout(opts SkyhookOperatorOptions, _package *v1alpha1.Packa
 		return _package.StageTimeout.Duration
 	}
 	return opts.JobStageTimeout
+}
+
+// createPodFromPackage creates a pod spec for a skyhook pod for a given package
+func createPodFromPackage(opts SkyhookOperatorOptions, _package *v1alpha1.Package, skyhook *wrapper.Skyhook, nodeName string, stage v1alpha1.Stage) *corev1.Pod {
+	// Generate consistent names that won't exceed k8s limits
+	volumeName := generateSafeName(63, "metadata", nodeName)
+	configMapName := generateSafeName(253, skyhook.Name, nodeName, "metadata")
+
+	volumes := []corev1.Volume{
+		{
+			Name: volumeNameRootMount,
+			VolumeSource: corev1.VolumeSource{
+				HostPath: &corev1.HostPathVolumeSource{
+					Path: "/",
+				},
+			},
+		},
+		{
+			Name: volumeName,
+			VolumeSource: corev1.VolumeSource{
+				ConfigMap: &corev1.ConfigMapVolumeSource{
+					LocalObjectReference: corev1.LocalObjectReference{
+						Name: configMapName,
+					},
+				},
+			},
+		},
+	}
+
+	volumeMounts := []corev1.VolumeMount{
+		{
+			Name:             volumeNameRootMount,
+			MountPath:        mountPathRoot,
+			MountPropagation: ptr(corev1.MountPropagationHostToContainer),
+		},
+		{
+			Name:      volumeName,
+			MountPath: "/skyhook-package/node-metadata",
+		},
+	}
+
+	if len(_package.ConfigMap) > 0 {
+		volumes = append(volumes, corev1.Volume{
+			Name: _package.Name,
+			VolumeSource: corev1.VolumeSource{
+				ConfigMap: &corev1.ConfigMapVolumeSource{
+					LocalObjectReference: corev1.LocalObjectReference{
+						Name: strings.ToLower(fmt.Sprintf("%s-%s-%s", skyhook.Name, _package.Name, _package.Version)),
+					},
+				},
+			},
+		})
+
+		// Mount each configMap key as its own subPath rather than mounting the
+		// whole configMap as a directory. A directory mount replaces the entire
+		// path, hiding any files the package image baked in under
+		// mountPathConfigMaps; per-key subPath mounts overlay individual files
+		// on top of the image content instead. Trade-off: subPath mounts do not
+		// receive live configMap updates, which is fine here because package
+		// pods are recreated per stage / version bump. Keys are iterated in
+		// sorted order so the generated pod spec is deterministic.
+		configMapKeys := make([]string, 0, len(_package.ConfigMap))
+		for key := range _package.ConfigMap {
+			configMapKeys = append(configMapKeys, key)
+		}
+		sort.Strings(configMapKeys)
+
+		for _, key := range configMapKeys {
+			volumeMounts = append(volumeMounts, corev1.VolumeMount{
+				Name:      _package.Name,
+				MountPath: fmt.Sprintf("%s/%s", mountPathConfigMaps, key),
+				SubPath:   key,
+				ReadOnly:  true,
+			})
+		}
+	}
+
+	copyDir := fmt.Sprintf("%s/%s/%s-%s-%s-%d",
+		opts.CopyDirRoot,
+		skyhook.Name,
+		_package.Name,
+		_package.Version,
+		skyhook.UID,
+		skyhook.Generation,
+	)
+	applyargs := []string{strings.ToLower(string(stage)), mountPathRoot, copyDir}
+	checkargs := []string{strings.ToLower(string(stage) + "-check"), mountPathRoot, copyDir}
+
+	agentEnvs := append(
+		_package.Env,
+		getAgentConfigEnvVars(opts, _package.Name, _package.Version, skyhook.ResourceID(), skyhook.Name, skyhook.NodeOrder(nodeName))...,
+	)
+
+	pod := &corev1.Pod{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      generateSafeName(63, skyhook.Name, _package.Name, _package.Version, string(stage), nodeName),
+			Namespace: opts.Namespace,
+			Labels: map[string]string{
+				fmt.Sprintf("%s/name", v1alpha1.METADATA_PREFIX):    skyhook.Name,
+				fmt.Sprintf("%s/package", v1alpha1.METADATA_PREFIX): fmt.Sprintf("%s-%s", _package.Name, _package.Version),
+			},
+		},
+		Spec: corev1.PodSpec{
+			NodeName:      nodeName,
+			RestartPolicy: corev1.RestartPolicyOnFailure,
+			InitContainers: []corev1.Container{
+				{
+					Name:            fmt.Sprintf("%s-init", trunstr(_package.Name, 43)),
+					Image:           getPackageImage(_package),
+					ImagePullPolicy: corev1.PullAlways,
+					Command:         []string{shellBinary},
+					Args: []string{
+						"-c",
+						"mkdir -p /root/${SKYHOOK_DIR} && cp -r /skyhook-package/* /root/${SKYHOOK_DIR}",
+					},
+					Env: []corev1.EnvVar{
+						{
+							Name:  "SKYHOOK_DIR",
+							Value: copyDir,
+						},
+					},
+					SecurityContext: &corev1.SecurityContext{
+						Privileged: ptr(true),
+					},
+					VolumeMounts: volumeMounts,
+				},
+				{
+					Name:            fmt.Sprintf("%s-%s", trunstr(_package.Name, 43), stage),
+					Image:           getAgentImage(opts, _package),
+					ImagePullPolicy: corev1.PullAlways,
+					Args:            applyargs,
+					Env:             agentEnvs,
+					SecurityContext: &corev1.SecurityContext{
+						Privileged: ptr(true),
+					},
+					VolumeMounts: volumeMounts,
+				},
+				{
+					Name:            fmt.Sprintf("%s-%scheck", trunstr(_package.Name, 43), stage),
+					Image:           getAgentImage(opts, _package),
+					ImagePullPolicy: corev1.PullAlways,
+					Args:            checkargs,
+					Env:             agentEnvs,
+					SecurityContext: &corev1.SecurityContext{
+						Privileged: ptr(true),
+					},
+					VolumeMounts: volumeMounts,
+				},
+			},
+			Containers: []corev1.Container{
+				{
+					Name:  pauseContainerName,
+					Image: opts.PauseImage,
+					Resources: corev1.ResourceRequirements{
+						Limits: corev1.ResourceList{
+							corev1.ResourceCPU:    resource.MustParse("100m"),
+							corev1.ResourceMemory: resource.MustParse("20Mi"),
+						},
+						Requests: corev1.ResourceList{
+							corev1.ResourceCPU:    resource.MustParse("100m"),
+							corev1.ResourceMemory: resource.MustParse("20Mi"),
+						},
+					},
+				},
+			},
+			Volumes:     volumes,
+			HostPID:     true,
+			HostNetwork: true,
+			// If you change these go change the SelectNode toleration in cluster_state.go
+			Tolerations: append([]corev1.Toleration{ // tolerate all cordon
+				{
+					Key:      TaintUnschedulable,
+					Operator: corev1.TolerationOpExists,
+				},
+				opts.GetRuntimeRequiredToleration(),
+			}, skyhook.Spec.AdditionalTolerations...),
+		},
+	}
+	if opts.ImagePullSecret != "" {
+		pod.Spec.ImagePullSecrets = []corev1.LocalObjectReference{
+			{
+				Name: opts.ImagePullSecret,
+			},
+		}
+	}
+	if _package.GracefulShutdown != nil {
+		pod.Spec.TerminationGracePeriodSeconds = ptr(int64(_package.GracefulShutdown.Duration.Seconds()))
+	}
+	setPodResources(pod, _package.Resources)
+	return pod
+}
+
+// createInterruptPodForPackage returns the pod spec for an interrupt pod given an package
+func createInterruptPodForPackage(opts SkyhookOperatorOptions, _interrupt *v1alpha1.Interrupt, argEncode string, _package *v1alpha1.Package, skyhook *wrapper.Skyhook, nodeName string, stage v1alpha1.Stage) *corev1.Pod {
+	copyDir := fmt.Sprintf("%s/%s/%s-%s-%s-%d",
+		opts.CopyDirRoot,
+		skyhook.Name,
+		_package.Name,
+		_package.Version,
+		skyhook.UID,
+		skyhook.Generation,
+	)
+
+	volumes := []corev1.Volume{
+		{
+			Name: volumeNameRootMount,
+			VolumeSource: corev1.VolumeSource{
+				HostPath: &corev1.HostPathVolumeSource{
+					Path: "/",
+				},
+			},
+		},
+		{
+			// node names in different CSPs might include dots which isn't allowed in volume names
+			// so we have to replace all dots with dashes
+			Name: generateSafeName(63, skyhook.Name, nodeName, "metadata"),
+			VolumeSource: corev1.VolumeSource{
+				ConfigMap: &corev1.ConfigMapVolumeSource{
+					LocalObjectReference: corev1.LocalObjectReference{
+						Name: strings.ReplaceAll(fmt.Sprintf("%s-%s-metadata", skyhook.Name, nodeName), ".", "-"),
+					},
+				},
+			},
+		},
+	}
+	volumeMounts := []corev1.VolumeMount{
+		{
+			Name:             volumeNameRootMount,
+			MountPath:        mountPathRoot,
+			MountPropagation: ptr(corev1.MountPropagationHostToContainer),
+		},
+	}
+
+	pod := &corev1.Pod{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      generateSafeName(63, skyhook.Name, string(stage), string(_interrupt.Type), nodeName),
+			Namespace: opts.Namespace,
+			Labels: map[string]string{
+				fmt.Sprintf("%s/name", v1alpha1.METADATA_PREFIX):      skyhook.Name,
+				fmt.Sprintf("%s/package", v1alpha1.METADATA_PREFIX):   fmt.Sprintf("%s-%s", _package.Name, _package.Version),
+				fmt.Sprintf("%s/interrupt", v1alpha1.METADATA_PREFIX): interruptLabelValue,
+			},
+		},
+		Spec: corev1.PodSpec{
+			NodeName:      nodeName,
+			RestartPolicy: corev1.RestartPolicyOnFailure,
+			InitContainers: []corev1.Container{
+				{
+					Name:  InterruptContainerName,
+					Image: getAgentImage(opts, _package),
+					Args:  []string{"interrupt", mountPathRoot, copyDir, argEncode},
+					Env:   getAgentConfigEnvVars(opts, _package.Name, _package.Version, skyhook.ResourceID(), skyhook.Name, skyhook.NodeOrder(nodeName)),
+					SecurityContext: &corev1.SecurityContext{
+						Privileged: ptr(true),
+					},
+					VolumeMounts: volumeMounts,
+					Resources: corev1.ResourceRequirements{
+						Limits: corev1.ResourceList{
+							corev1.ResourceCPU:    resource.MustParse("500m"),
+							corev1.ResourceMemory: resource.MustParse("64Mi"),
+						},
+						Requests: corev1.ResourceList{
+							corev1.ResourceCPU:    resource.MustParse("500m"),
+							corev1.ResourceMemory: resource.MustParse("64Mi"),
+						},
+					},
+				},
+			},
+			Containers: []corev1.Container{
+				{
+					Name:  pauseContainerName,
+					Image: opts.PauseImage,
+					Resources: corev1.ResourceRequirements{
+						Limits: corev1.ResourceList{
+							corev1.ResourceCPU:    resource.MustParse("100m"),
+							corev1.ResourceMemory: resource.MustParse("20Mi"),
+						},
+						Requests: corev1.ResourceList{
+							corev1.ResourceCPU:    resource.MustParse("100m"),
+							corev1.ResourceMemory: resource.MustParse("20Mi"),
+						},
+					},
+				},
+			},
+			HostPID:     true,
+			HostNetwork: true,
+			// If you change these go change the SelectNode toleration in cluster_state.go
+			Tolerations: append([]corev1.Toleration{ // tolerate all cordon
+				{
+					Key:      TaintUnschedulable,
+					Operator: corev1.TolerationOpExists,
+				},
+				opts.GetRuntimeRequiredToleration(),
+			}, skyhook.Spec.AdditionalTolerations...),
+			Volumes: volumes,
+		},
+	}
+	if opts.ImagePullSecret != "" {
+		pod.Spec.ImagePullSecrets = []corev1.LocalObjectReference{
+			{
+				Name: opts.ImagePullSecret,
+			},
+		}
+	}
+	return pod
+}
+
+// PodMatchesPackage asserts that a given pod matches the given pod spec
+func podMatchesPackage(opts SkyhookOperatorOptions, _package *v1alpha1.Package, pod corev1.Pod, skyhook *wrapper.Skyhook, stage v1alpha1.Stage) bool {
+	var expectedPod *corev1.Pod
+
+	// need to differentiate whether the pod is for an interrupt or not so we know
+	// what to expect and how to compare them
+	isInterrupt := false
+	_, limitRange := pod.Annotations["kubernetes.io/limit-ranger"]
+
+	if pod.Labels[fmt.Sprintf("%s/interrupt", v1alpha1.METADATA_PREFIX)] == interruptLabelValue {
+		expectedPod = createInterruptPodForPackage(opts, &v1alpha1.Interrupt{}, "", _package, skyhook, "", stage)
+		isInterrupt = true
+	} else {
+		expectedPod = createPodFromPackage(opts, _package, skyhook, "", stage)
+	}
+
+	actualPod := pod.DeepCopy()
+
+	// check to see whether the name or the version of the package changed
+	packageLabel := fmt.Sprintf("%s/package", v1alpha1.METADATA_PREFIX)
+	if actualPod.Labels[packageLabel] != expectedPod.Labels[packageLabel] {
+		return false
+	}
+
+	// compare initContainers since this is where a lot of the important info lives
+	for i := range actualPod.Spec.InitContainers {
+		expectedContainer := expectedPod.Spec.InitContainers[i]
+		actualContainer := actualPod.Spec.InitContainers[i]
+
+		if expectedContainer.Name != actualContainer.Name {
+			return false
+		}
+
+		if expectedContainer.Image != actualContainer.Image {
+			return false
+		}
+
+		// compare the containers env vars except for the ones that are inserted
+		// by the operator by default as the SKYHOOK_RESOURCE_ID will change every
+		// time the skyhook is updated and would cause every pod to be removed
+		// TODO: This is ignoring all the static env vars that are set by operator config.
+		// It probably should be just SKYHOOK_RESOURCE_ID that is ignored. Otherwise,
+		// a user will have to manually delete the pod to update the package when operator is updated.
+		dummyAgentEnv := getAgentConfigEnvVars(opts, "", "", "", "", 0)
+		excludedEnvs := make([]string, len(dummyAgentEnv))
+		for i, env := range dummyAgentEnv {
+			excludedEnvs[i] = env.Name
+		}
+		expectedFilteredEnv := FilterEnv(expectedContainer.Env, excludedEnvs...)
+		actualFilteredEnv := FilterEnv(actualContainer.Env, excludedEnvs...)
+		if !reflect.DeepEqual(expectedFilteredEnv, actualFilteredEnv) {
+			return false
+		}
+
+		if !isInterrupt { // dont compare these since they are not configured on interrupt
+			// compare resource requests and limits (CPU, memory, etc.)
+			expectedResources := expectedContainer.Resources
+			actualResources := actualContainer.Resources
+			if skyhook.Spec.Packages[_package.Name].Resources != nil {
+				// If CR has resources specified, they should match exactly
+				if !reflect.DeepEqual(expectedResources, actualResources) {
+					return false
+				}
+			} else {
+				// If CR has no resources specified, ensure pod has no resource overrides
+				if !limitRange {
+					if actualResources.Requests != nil || actualResources.Limits != nil {
+						return false
+					}
+				}
+			}
+		}
+	}
+
+	return true
 }
