+++ operator/internal/controller/skyhook_controller_test.go
@@ -1322,144 +1322,6 @@ var _ = Describe("skyhook controller tests", func() {
 		Expect(toleration.ToleratesTaint(logger, &taint, false)).To(BeTrue())
 
 	})
-	It("Pods should always tolerate runtime required taint", func() {
-		pod := createPodFromPackage(
-			operator.opts,
-			&v1alpha1.Package{
-				PackageRef: v1alpha1.PackageRef{
-					Name:    "foo",
-					Version: "1.1.2",
-				},
-				Image: "foo/bar",
-			},
-			&wrapper.Skyhook{
-				NodeWright: &v1alpha1.NodeWright{
-					Spec: v1alpha1.NodeWrightSpec{
-						RuntimeRequired: true,
-					},
-				},
-			},
-			"node1",
-			v1alpha1.StageApply,
-		)
-		found_toleration := false
-		expected_toleration := opts.GetRuntimeRequiredToleration()
-		for _, toleration := range pod.Spec.Tolerations {
-			if toleration.Key == expected_toleration.Key && toleration.Value == expected_toleration.Value && toleration.Effect == expected_toleration.Effect {
-				found_toleration = true
-				break
-			}
-		}
-		Expect(found_toleration).To(BeTrue())
-	})
-	It("Interrupt pods should tolerate runtime required taint when it is runtime required", func() {
-		pod := createInterruptPodForPackage(
-			operator.opts,
-			&v1alpha1.Interrupt{
-				Type: v1alpha1.REBOOT,
-			},
-			"argEncode",
-
-			&v1alpha1.Package{
-				PackageRef: v1alpha1.PackageRef{
-					Name:    "foo",
-					Version: "1.1.2",
-				},
-				Image: "foo/bar",
-			},
-			&wrapper.Skyhook{
-				NodeWright: &v1alpha1.NodeWright{
-					Spec: v1alpha1.NodeWrightSpec{
-						RuntimeRequired: true,
-					},
-				},
-			},
-			"node1",
-			v1alpha1.StageInterrupt,
-		)
-		found_toleration := false
-		expected_toleration := opts.GetRuntimeRequiredToleration()
-		for _, toleration := range pod.Spec.Tolerations {
-			if toleration.Key == expected_toleration.Key && toleration.Value == expected_toleration.Value && toleration.Effect == expected_toleration.Effect {
-				found_toleration = true
-				break
-			}
-		}
-		Expect(found_toleration).To(BeTrue())
-	})
-
-	It("Pods should not have imagePullSecrets when ImagePullSecret is empty", func() {
-		emptyOpts := SkyhookOperatorOptions{
-			Namespace:            "skyhook",
-			MaxInterval:          time.Second * 61,
-			ImagePullSecret:      "", // Empty - no pull secret
-			CopyDirRoot:          "/tmp",
-			ReapplyOnReboot:      true,
-			RuntimeRequiredTaint: "skyhook.nvidia.com=runtime-required:NoSchedule",
-			AgentImage:           "foo:bar",
-			PauseImage:           "foo:bar",
-		}
-
-		pod := createPodFromPackage(
-			emptyOpts,
-			&v1alpha1.Package{
-				PackageRef: v1alpha1.PackageRef{
-					Name:    "foo",
-					Version: "1.1.2",
-				},
-				Image: "foo/bar",
-			},
-			&wrapper.Skyhook{
-				NodeWright: &v1alpha1.NodeWright{
-					Spec: v1alpha1.NodeWrightSpec{
-						RuntimeRequired: true,
-					},
-				},
-			},
-			"node1",
-			v1alpha1.StageApply,
-		)
-		Expect(pod.Spec.ImagePullSecrets).To(BeEmpty())
-	})
-
-	It("Interrupt pods should not have imagePullSecrets when ImagePullSecret is empty", func() {
-		emptyOpts := SkyhookOperatorOptions{
-			Namespace:            "skyhook",
-			MaxInterval:          time.Second * 61,
-			ImagePullSecret:      "", // Empty - no pull secret
-			CopyDirRoot:          "/tmp",
-			ReapplyOnReboot:      true,
-			RuntimeRequiredTaint: "skyhook.nvidia.com=runtime-required:NoSchedule",
-			AgentImage:           "foo:bar",
-			PauseImage:           "foo:bar",
-		}
-
-		pod := createInterruptPodForPackage(
-			emptyOpts,
-			&v1alpha1.Interrupt{
-				Type: v1alpha1.REBOOT,
-			},
-			"argEncode",
-			&v1alpha1.Package{
-				PackageRef: v1alpha1.PackageRef{
-					Name:    "foo",
-					Version: "1.1.2",
-				},
-				Image: "foo/bar",
-			},
-			&wrapper.Skyhook{
-				NodeWright: &v1alpha1.NodeWright{
-					Spec: v1alpha1.NodeWrightSpec{
-						RuntimeRequired: true,
-					},
-				},
-			},
-			"node1",
-			v1alpha1.StageInterrupt,
-		)
-		Expect(pod.Spec.ImagePullSecrets).To(BeEmpty())
-	})
-
 	It("should generate deterministic pod names", func() {
 		// Setup basic test data
 		skyhook := &wrapper.Skyhook{
@@ -1534,78 +1396,6 @@ var _ = Describe("skyhook controller tests", func() {
 		Expect(len(longName)).To(BeNumerically("<=", 63), "Pod name should not exceed Kubernetes 63 character limit")
 		Expect(longName).To(MatchRegexp(`-[0-9a-f]+$`), "Pod name should end with a hash component")
 	})
-
-	It("should correctly identify if a pod matches a package", func() {
-
-		// Create a test package
-		testPackage := &v1alpha1.Package{
-			PackageRef: v1alpha1.PackageRef{
-				Name:    "test-package",
-				Version: "1.2.3",
-			},
-			Image: "test-image:1.2.3",
-			Resources: &v1alpha1.ResourceRequirements{
-				CPURequest:    resource.MustParse("100m"),
-				CPULimit:      resource.MustParse("200m"),
-				MemoryRequest: resource.MustParse("64Mi"),
-				MemoryLimit:   resource.MustParse("128Mi"),
-			},
-		}
-
-		// Create a test skyhook
-		testSkyhook := &wrapper.Skyhook{
-			NodeWright: &v1alpha1.NodeWright{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "test-skyhook",
-				},
-				Spec: v1alpha1.NodeWrightSpec{
-					Packages: v1alpha1.Packages{
-						"test-package": *testPackage,
-					},
-				},
-			},
-		}
-
-		// Stage to test
-		testStage := v1alpha1.StageApply
-
-		// Create actual pods that would be created by the operator functions
-		// First using CreatePodFromPackage
-		actualPod := createPodFromPackage(operator.opts, testPackage, testSkyhook, "test-node", testStage)
-
-		// Verify that the pod matches the package according to PodMatchesPackage
-		matches := podMatchesPackage(operator.opts, testPackage, *actualPod, testSkyhook, testStage)
-		Expect(matches).To(BeTrue(), "PodMatchesPackage should recognize the pod it created")
-
-		// Now let's modify the package version and see if it correctly identifies non-matches
-		modifiedPackage := testPackage.DeepCopy()
-		modifiedPackage.Version = "1.2.4"
-
-		matches = podMatchesPackage(operator.opts, modifiedPackage, *actualPod, testSkyhook, testStage)
-		Expect(matches).To(BeFalse(), "PodMatchesPackage should not match when package version changed")
-
-		// Test with different stage
-		matches = podMatchesPackage(operator.opts, testPackage, *actualPod, testSkyhook, v1alpha1.StageConfig)
-		Expect(matches).To(BeFalse(), "PodMatchesPackage should not match when stage changed")
-
-		// Test with interrupt pods
-		interruptPod := createInterruptPodForPackage(
-			operator.opts,
-			&v1alpha1.Interrupt{
-				Type: v1alpha1.REBOOT,
-			},
-			"argEncode",
-			testPackage,
-			testSkyhook,
-			"test-node",
-			testStage,
-		)
-
-		// Verify that the interrupt pod matches the package
-		matches = podMatchesPackage(operator.opts, testPackage, *interruptPod, testSkyhook, testStage)
-		Expect(matches).To(BeTrue(), "PodMatchesPackage should recognize the interrupt pod it created")
-	})
-
 	It("should generate valid volume names", func() {
 		tests := []struct {
 			name        string
@@ -1651,78 +1441,6 @@ var _ = Describe("skyhook controller tests", func() {
 			Expect(result).To(MatchRegexp(`^[a-z0-9]([-a-z0-9]*[a-z0-9])?$`), "volume name should match kubernetes naming requirements")
 		}
 	})
-
-	It("should mount each configMap key as a subPath so image defaults are not clobbered", func() {
-		testPackage := &v1alpha1.Package{
-			PackageRef: v1alpha1.PackageRef{
-				Name:    "foo",
-				Version: "1.1.2",
-			},
-			Image: "foo/bar",
-			ConfigMap: map[string]string{
-				"b.sh": "echo b",
-				"a.sh": "echo a",
-				"c.sh": "echo c",
-			},
-		}
-		testSkyhook := &wrapper.Skyhook{
-			NodeWright: &v1alpha1.NodeWright{
-				ObjectMeta: metav1.ObjectMeta{Name: "test-skyhook"},
-			},
-		}
-
-		pod := createPodFromPackage(operator.opts, testPackage, testSkyhook, "node1", v1alpha1.StageApply)
-
-		container := pod.Spec.InitContainers[0]
-
-		// The configMap must never be mounted as a bare directory: that hides
-		// any files the package image baked in at /skyhook-package/configmaps.
-		for _, vm := range container.VolumeMounts {
-			if vm.MountPath == "/skyhook-package/configmaps" {
-				Fail("configMap must not be mounted as a directory; expected per-key subPath mounts")
-			}
-		}
-
-		// Collect the configMap mounts (those backed by the package volume).
-		configMapMounts := make([]corev1.VolumeMount, 0)
-		for _, vm := range container.VolumeMounts {
-			if vm.Name == testPackage.Name {
-				configMapMounts = append(configMapMounts, vm)
-			}
-		}
-
-		Expect(configMapMounts).To(HaveLen(3))
-		// Sorted-key order for a deterministic pod spec.
-		Expect(configMapMounts[0]).To(Equal(corev1.VolumeMount{
-			Name:      testPackage.Name,
-			MountPath: "/skyhook-package/configmaps/a.sh",
-			SubPath:   "a.sh",
-			ReadOnly:  true,
-		}))
-		Expect(configMapMounts[1]).To(Equal(corev1.VolumeMount{
-			Name:      testPackage.Name,
-			MountPath: "/skyhook-package/configmaps/b.sh",
-			SubPath:   "b.sh",
-			ReadOnly:  true,
-		}))
-		Expect(configMapMounts[2]).To(Equal(corev1.VolumeMount{
-			Name:      testPackage.Name,
-			MountPath: "/skyhook-package/configmaps/c.sh",
-			SubPath:   "c.sh",
-			ReadOnly:  true,
-		}))
-
-		// The underlying configMap volume is still mounted exactly once.
-		configMapVolumes := 0
-		for _, v := range pod.Spec.Volumes {
-			if v.Name == testPackage.Name {
-				configMapVolumes++
-				Expect(v.ConfigMap).NotTo(BeNil())
-			}
-		}
-		Expect(configMapVolumes).To(Equal(1))
-	})
-
 	It("should generate valid configmap names", func() {
 		tests := []struct {
 			name        string
@@ -1922,239 +1640,7 @@ var _ = Describe("skyhook controller tests", func() {
 	})
 })
 
-var _ = Describe("Resource Comparison", func() {
-	var (
-		expectedPod *corev1.Pod
-		actualPod   *corev1.Pod
-		skyhook     *wrapper.Skyhook
-		package_    *v1alpha1.Package
-	)
-
-	BeforeEach(func() {
-		// Setup common test objects
-		nodeName := "testNode"
-		stage := v1alpha1.StageApply
-		package_ = &v1alpha1.Package{
-			PackageRef: v1alpha1.PackageRef{
-				Name:    "test-package",
-				Version: "1.0.0",
-			},
-			Image: "test-image",
-		}
-
-		skyhook = &wrapper.Skyhook{
-			NodeWright: &v1alpha1.NodeWright{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "test-skyhook",
-				},
-				Spec: v1alpha1.NodeWrightSpec{
-					Packages: map[string]v1alpha1.Package{
-						"test-package": *package_,
-					},
-				},
-			},
-		}
-
-		// Create base pod structure, to much work to do it again
-		expectedPod = createPodFromPackage(operator.opts, package_, skyhook, nodeName, stage)
-		actualPod = expectedPod.DeepCopy()
-	})
-
-	It("should match when resources are identical", func() {
-		// Setup: Add resources to package and expected pod
-		newPackage := *package_
-		newPackage.Resources = &v1alpha1.ResourceRequirements{
-			CPURequest:    resource.MustParse("100m"),
-			CPULimit:      resource.MustParse("200m"),
-			MemoryRequest: resource.MustParse("128Mi"),
-			MemoryLimit:   resource.MustParse("256Mi"),
-		}
-		skyhook.Spec.Packages["test-package"] = newPackage
-
-		expectedResources := corev1.ResourceRequirements{
-			Requests: corev1.ResourceList{
-				corev1.ResourceCPU:    resource.MustParse("100m"),
-				corev1.ResourceMemory: resource.MustParse("128Mi"),
-			},
-			Limits: corev1.ResourceList{
-				corev1.ResourceCPU:    resource.MustParse("200m"),
-				corev1.ResourceMemory: resource.MustParse("256Mi"),
-			},
-		}
-
-		// Set resources for all init containers in expected pod
-		for i := range expectedPod.Spec.InitContainers {
-			expectedPod.Spec.InitContainers[i].Resources = expectedResources
-		}
-
-		// Test: Set actual pod resources to match expected
-		for i := range actualPod.Spec.InitContainers {
-			actualPod.Spec.InitContainers[i].Resources = expectedResources
-		}
-
-		// Set the package in the pod annotations
-		err := SetPackages(actualPod, skyhook.NodeWright, newPackage.Image, v1alpha1.StageApply, &newPackage)
-		Expect(err).ToNot(HaveOccurred())
-
-		Expect(podMatchesPackage(operator.opts, &newPackage, *actualPod, skyhook, v1alpha1.StageApply)).To(BeTrue())
-	})
-
-	It("should not match when resources differ", func() {
-		// Setup: Add resources to package and expected pod
-		newPackage := *package_
-		newPackage.Resources = &v1alpha1.ResourceRequirements{
-			CPURequest:    resource.MustParse("100m"),
-			CPULimit:      resource.MustParse("200m"),
-			MemoryRequest: resource.MustParse("128Mi"),
-			MemoryLimit:   resource.MustParse("256Mi"),
-		}
-		skyhook.Spec.Packages["test-package"] = newPackage
-
-		expectedResources := corev1.ResourceRequirements{
-			Requests: corev1.ResourceList{
-				corev1.ResourceCPU:    resource.MustParse("100m"),
-				corev1.ResourceMemory: resource.MustParse("128Mi"),
-			},
-			Limits: corev1.ResourceList{
-				corev1.ResourceCPU:    resource.MustParse("200m"),
-				corev1.ResourceMemory: resource.MustParse("256Mi"),
-			},
-		}
-
-		// Set resources for all init containers in expected pod
-		for i := range expectedPod.Spec.InitContainers {
-			expectedPod.Spec.InitContainers[i].Resources = expectedResources
-		}
-
-		// Test: Set different CPU request in actual pod for all init containers
-		differentResources := corev1.ResourceRequirements{
-			Requests: corev1.ResourceList{
-				corev1.ResourceCPU:    resource.MustParse("200m"), // Different CPU request
-				corev1.ResourceMemory: resource.MustParse("128Mi"),
-			},
-			Limits: corev1.ResourceList{
-				corev1.ResourceCPU:    resource.MustParse("200m"),
-				corev1.ResourceMemory: resource.MustParse("256Mi"),
-			},
-		}
-		for i := range actualPod.Spec.InitContainers {
-			actualPod.Spec.InitContainers[i].Resources = differentResources
-		}
-
-		// Set the package in the pod annotations
-		err := SetPackages(actualPod, skyhook.NodeWright, newPackage.Image, v1alpha1.StageApply, &newPackage)
-		Expect(err).ToNot(HaveOccurred())
-
-		Expect(podMatchesPackage(operator.opts, &newPackage, *actualPod, skyhook, v1alpha1.StageApply)).To(BeFalse())
-	})
-
-	It("should match when no resources are specified and pod has no overrides", func() {
-		// Setup: Ensure no resources in package
-		newPackage := *package_
-		newPackage.Resources = nil
-		skyhook.Spec.Packages["test-package"] = newPackage
-
-		// Test: Ensure pod has no resource overrides for any init container
-		emptyResources := corev1.ResourceRequirements{}
-		for i := range actualPod.Spec.InitContainers {
-			actualPod.Spec.InitContainers[i].Resources = emptyResources
-		}
-
-		// Set the package in the pod annotations
-		err := SetPackages(actualPod, skyhook.NodeWright, newPackage.Image, v1alpha1.StageApply, &newPackage)
-		Expect(err).ToNot(HaveOccurred())
-
-		Expect(podMatchesPackage(operator.opts, &newPackage, *actualPod, skyhook, v1alpha1.StageApply)).To(BeTrue())
-	})
-
-	It("should not match when no resources are specified but pod has requests", func() {
-		// Setup: Ensure no resources in package
-		newPackage := *package_
-		newPackage.Resources = nil
-		skyhook.Spec.Packages["test-package"] = newPackage
-
-		// Test: Add resource requests to all init containers
-		requestResources := corev1.ResourceRequirements{
-			Requests: corev1.ResourceList{
-				corev1.ResourceCPU:    resource.MustParse("100m"),
-				corev1.ResourceMemory: resource.MustParse("128Mi"),
-			},
-		}
-		for i := range actualPod.Spec.InitContainers {
-			actualPod.Spec.InitContainers[i].Resources = requestResources
-		}
-
-		// Set the package in the pod annotations
-		err := SetPackages(actualPod, skyhook.NodeWright, newPackage.Image, v1alpha1.StageApply, &newPackage)
-		Expect(err).ToNot(HaveOccurred())
-
-		Expect(podMatchesPackage(operator.opts, &newPackage, *actualPod, skyhook, v1alpha1.StageApply)).To(BeFalse())
-	})
-
-	It("should not match when no resources are specified but pod has limits", func() {
-		// Setup: Ensure no resources in package
-		newPackage := *package_
-		newPackage.Resources = nil
-		skyhook.Spec.Packages["test-package"] = newPackage
-
-		// Test: Add resource limits to all init containers
-		limitResources := corev1.ResourceRequirements{
-			Limits: corev1.ResourceList{
-				corev1.ResourceCPU:    resource.MustParse("200m"),
-				corev1.ResourceMemory: resource.MustParse("256Mi"),
-			},
-		}
-		for i := range actualPod.Spec.InitContainers {
-			actualPod.Spec.InitContainers[i].Resources = limitResources
-		}
-
-		// Set the package in the pod annotations
-		err := SetPackages(actualPod, skyhook.NodeWright, newPackage.Image, v1alpha1.StageApply, &newPackage)
-		Expect(err).ToNot(HaveOccurred())
-
-		Expect(podMatchesPackage(operator.opts, &newPackage, *actualPod, skyhook, v1alpha1.StageApply)).To(BeFalse())
-	})
-
-	It("should ignore SKYHOOK_RESOURCE_ID env var", func() {
-		newPackage := *package_
-		newPackage.Resources = nil
-		skyhook.Spec.Packages["test-package"] = newPackage
-
-		// Setup: Add SKYHOOK_RESOURCE_ID env var to all init containers
-		for i := range actualPod.Spec.InitContainers {
-			actualPod.Spec.InitContainers[i].Env = append(actualPod.Spec.InitContainers[i].Env, corev1.EnvVar{
-				Name:  "SKYHOOK_RESOURCE_ID",
-				Value: "SOME_VALUE",
-			})
-		}
-
-		// Set the package in the pod annotations
-		err := SetPackages(actualPod, skyhook.NodeWright, newPackage.Image, v1alpha1.StageApply, &newPackage)
-		Expect(err).ToNot(HaveOccurred())
-
-		Expect(podMatchesPackage(operator.opts, &newPackage, *actualPod, skyhook, v1alpha1.StageApply)).To(BeTrue())
-	})
-
-	It("should not ignore non static env vars", func() {
-		newPackage := *package_
-		newPackage.Resources = nil
-		skyhook.Spec.Packages["test-package"] = newPackage
-
-		// Setup: Add SKYHOOK_RESOURCE_ID env var to all init containers
-		for i := range actualPod.Spec.InitContainers {
-			actualPod.Spec.InitContainers[i].Env = append(actualPod.Spec.InitContainers[i].Env, corev1.EnvVar{
-				Name:  "SOME_ENV_VAR",
-				Value: "SOME_VALUE",
-			})
-		}
-
-		// Set the package in the pod annotations
-		err := SetPackages(actualPod, skyhook.NodeWright, newPackage.Image, v1alpha1.StageApply, &newPackage)
-		Expect(err).ToNot(HaveOccurred())
-
-		Expect(podMatchesPackage(operator.opts, &newPackage, *actualPod, skyhook, v1alpha1.StageApply)).To(BeFalse())
-	})
-
+var _ = Describe("cluster state compartments", func() {
 	It("should partition nodes into compartments", func() {
 		skyhooks := &v1alpha1.NodeWrightList{
 			Items: []v1alpha1.NodeWright{
